requirements.txt
discord.py>=2.3.0
requests>=2.31.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
pytz>=2024.1
lxml>=4.9.0
